rename cli.main to cli.build_cli
